social-media-agent/requirements.txt
anthropic>=0.86.0
google-genai>=1.0.0
Pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0
rich>=13.0.0
pytest>=8.0.0
